wsl not viable, qemu on wx64 is not an option
qemu with whpx + virgl opengl on windows has been achieved, but arm translation is broken. i suspect it has something to do with binfmt_misc. when I was looking through many android distros, specifically one that actually had arm64 translation, none of them worked, until i tried the current blissos with binfmt_misc installed, it actually worked, but yeah, idk bout windows, and ive currently suspended development of wx86_64, as yes, ive made a better android experience even on windows, but not for sideswipe. which is what my goal is. so, not releasing. ty for reading if you somehow did read this